ref: move create_msg functions to a new file in the common folder
